+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/Cryptography/LocalCryptographyProvider.cs
@@ -9,13 +9,13 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
 {
     internal abstract class LocalCryptographyProvider : ICryptographyProvider
     {
-        private readonly Key _key;
+        private readonly KeyVaultKey _key;
 
-        public LocalCryptographyProvider(Key key)
+        public LocalCryptographyProvider(KeyVaultKey key)
         {
             _key = key ?? throw new ArgumentNullException(nameof(key));
 
-            KeyMaterial = key.KeyMaterial;
+            KeyMaterial = key.Key;
         }
 
         public bool ShouldRemote => KeyMaterial?.Id != null;
@@ -100,9 +100,9 @@ namespace Azure.Security.KeyVault.Keys.Cryptography
                 throw new InvalidOperationException($"The key \"{_key.Name}\" is not valid before {_key.Properties.NotBefore.Value:r}.");
             }
 
-            if (_key.Properties.Expires.HasValue && now > _key.Properties.Expires.Value)
+            if (_key.Properties.ExpiresOn.HasValue && now > _key.Properties.ExpiresOn.Value)
             {
-                throw new InvalidOperationException($"The key \"{_key.Name}\" is not valid after {_key.Properties.Expires.Value:r}.");
+                throw new InvalidOperationException($"The key \"{_key.Name}\" is not valid after {_key.Properties.ExpiresOn.Value:r}.");
             }
         }
     }