+++ sdk/keyvault/Azure.Security.KeyVault.Keys/src/KeyClientOptions.cs
@@ -41,7 +41,7 @@ namespace Azure.Security.KeyVault.Keys
         public ServiceVersion Version { get; }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="KeyClientOptions"/>
+        /// Initializes a new instance of the <see cref="KeyClientOptions"/> class.
         /// class.
         /// </summary>
         /// <param name="version">